updated style with vanilla css

+++ backend/app.py
@@ -15,7 +15,7 @@ class Team(db.Model):
     games_played = db.Column(db.Integer)
     total_wins = db.Column(db.Integer)
     total_losses = db.Column(db.Integer)
-    win_percentage = db.Column(db.Float(3))
+    win_percentage = db.Column(db.Float(precision=1))
     streak = db.Column(db.String(50))
